Breaks markup into components for Header, Player, and Counter
The application was broken into sub React components using the same
function return style as the Application -- each now takes a default
"props" argument which refers to the React props object (which you
can set using the React.propTypes and React.defaultProps built-in
objects).

The application is now two custom tags - for header and player -
and the player component contains the custom tag for the counter.
This breaks the single large app into reusable components.