Moves players into a player array, creates components from array
JSX can't have expressions in it, only statements, so a for loop
won't work. Instead, you can use a map array method to iterate over
the player array and return components using the data.

Iteration in React requires having a unique "key" attribute on your
data so it can more efficiently re-render only what is necessary
when changes are made to the data.